+++ ICD.Connect.Devices/Telemetry/DeviceInfo/Monitored/MonitoredNetworkDeviceInfo.cs
@@ -4,7 +4,7 @@ namespace ICD.Connect.Devices.Telemetry.DeviceInfo.Monitored
 {
 	public sealed class MonitoredNetworkDeviceInfo : AbstractNetworkDeviceInfo<MonitoredAdapterNetworkDeviceInfo>, IMonitoredNetworkDeviceInfo
 	{
-		protected override MonitoredAdapterNetworkDeviceInfo CreateNewAdapter(int address)
+		protected override IAdapterNetworkDeviceInfo CreateNewAdapter(int address)
 		{
 			return new MonitoredAdapterNetworkDeviceInfo(address);
 		}